Remove CUI to OMIM mapper
Remove the CUI to OMIM id resolver using the UMLS API. This was
intended to provide greater id mapping coverage from Orphanet.
However, it is slow and does not result in much useable information,
which is why it is being removed.

Most of the OMIM ids returned are MTHU ids, which I believe are
deprecated and therefore useless when trying to compare to the
normal six digit OMIM ids used to cross reference Orphanet
diseases.